+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Validation of parameter based on the RPC Type
+
+
 ## [0.6.2] - 2025-02-04
 ### Changed
 - Minimal supported version of pySHV is now 0.8.0